DEV-2 refactoring
