+++ src/Login/Login.jsx
@@ -1,24 +1,10 @@
 import React, {useContext} from 'react'
-import {
-  Box,
-  makeStyles,
-  Paper,
-} from "@material-ui/core";
 import {LoginModal} from "./LoginModal/LoginModal";
-import {LogoForHeader} from "./LogoForHeader/LogoForHeader";
 import {AuthContext} from "../AuthContext";
 import PropTypes from "prop-types";
-
-const useStyles = makeStyles(() => ({
-  backgroundBlack: {
-    backgroundColor: "black",
-    color: "white"
-  },
-}));
-
+import {BasePage} from "../helpers/BasePage";
 
 export const Login = ({navigateTo}) => {
-  const classes = useStyles();
   const {logIn} = useContext(AuthContext)
   const onSubmitLogin = (event) => {
     event.preventDefault()
@@ -26,19 +12,10 @@ export const Login = ({navigateTo}) => {
     logIn(email.value, password.value)
   }
   return (
-    <Box display="flex" direction="row">
-      <Box width="34%" height='100vh' display="flex" className={classes.backgroundBlack} alignItems='center'
-           justifyContent="center">
-        <LogoForHeader/>
-      </Box>
-      <Box height='100vh' width="66%" display="flex" alignItems='center' justifyContent="center">
-        <Paper elevation={5} style={{borderRadius: "20px"}}>
-          <Box width="520px" padding='48px 0'>
-            <LoginModal navigateTo={navigateTo} onSubmit={onSubmitLogin}/>
-          </Box>
-        </Paper>
-      </Box>
-    </Box>
+    <BasePage>
+      <LoginModal navigateTo={navigateTo} onSubmit={onSubmitLogin}/>
+    </BasePage>
+
   )
 }
