+++ src/SingUp/SingUp.jsx
@@ -1,37 +1,14 @@
 import React from 'react'
-import {
-  Box,
-  makeStyles,
-  Paper,
-} from "@material-ui/core";
 import {SingUpModal} from "./SingUpModal/SingUpModal";
-import {LogoForHeader} from "../Login/LogoForHeader/LogoForHeader";
 import PropTypes from "prop-types";
-
-const useStyles = makeStyles(() => ({
-  backgroundBlack: {
-    backgroundColor: "black",
-    color: "white"
-  },
-}));
-
+import {BasePage} from "../helpers/BasePage";
 
 export const SingUp = ({navigateTo}) => {
-  const classes = useStyles();
   return (
-    <Box display="flex" direction="row">
-      <Box width="34%" height='100vh' display="flex" className={classes.backgroundBlack} alignItems='center'
-           justifyContent="center">
-        <LogoForHeader/>
-      </Box>
-      <Box height='100vh' width="66%" display="flex" alignItems='center' justifyContent="center">
-        <Paper elevation={5} style={{borderRadius: "20px"}}>
-          <Box width="520px" padding='48px 0'>
-            <SingUpModal navigateTo={navigateTo}/>
-          </Box>
-        </Paper>
-      </Box>
-    </Box>
+    <BasePage>
+      <SingUpModal navigateTo={navigateTo}/>
+    </BasePage>
+
   )
 }
